requirements.txt
# Core dependencies
streamlit>=1.28.0
anthropic>=0.3.0
pandas>=2.0.0
numpy>=1.24.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Vector operations
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
streamlit-authenticator>=0.2.0